pyproject.toml
[project]
name = "ecommerce-faker-data-generator"
version = "0.1.0"
description = ""
authors = [
    {name = "Bao Nguyen ",email = "baronlq117@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faker (>=39.0.0,<40.0.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
